+++ pkg/analyzer/lib/dart/element/scope.dart
@@ -9,5 +9,20 @@ import 'package:meta/meta.dart';
 abstract class Scope {
   /// Return the element with the name `id` or `id=` (if [setter] is `true`),
   /// `null` if the name is not defined within this scope.
+  @Deprecated('Use lookup2() that is closer to the language specification')
   Element lookup({@required String id, @required bool setter});
+
+  /// Return the result of lexical lookup for the given [id], not `null`.
+  ///
+  /// Getters and setters are bundled, when we found one or another, we are
+  /// done with the lookup, and return both the getter and the setter, if
+  /// available.
+  ScopeLookupResult lookup2(String id);
+}
+
+class ScopeLookupResult {
+  final Element getter;
+  final Element setter;
+
+  ScopeLookupResult(this.getter, this.setter);
 }