+++ pkg/analyzer/test/src/dart/resolution/library_element_test.dart
@@ -123,7 +123,7 @@ class A {}
     await _assertLanguageVersionCurrent('package:aaa/c.dart');
   }
 
-  test_scope() async {
+  test_scope_lookup2() async {
     await assertNoErrorsInCode(r'''
 int foo;
 ''');
@@ -131,41 +131,41 @@ int foo;
     var scope = result.libraryElement.scope;
 
     assertElement(
-      scope.lookup(id: 'foo', setter: false),
+      scope.lookup2('foo').getter,
       findElement.topGet('foo'),
     );
     assertElement(
-      scope.lookup(id: 'foo', setter: true),
+      scope.lookup2('foo').setter,
       findElement.topSet('foo'),
     );
   }
 
-  test_scope_implicitCoreImport() async {
+  test_scope_lookup2__implicitCoreImport() async {
     await assertNoErrorsInCode('');
 
     var scope = result.libraryElement.scope;
 
     assertElement(
-      scope.lookup(id: 'int', setter: false),
+      scope.lookup2('int').getter,
       intElement,
     );
   }
 
-  test_scope_notFound() async {
+  test_scope_lookup2__notFound() async {
     await assertNoErrorsInCode('');
 
     var scope = result.libraryElement.scope;
 
     assertElementNull(
-      scope.lookup(id: 'noSuchGetter', setter: false),
+      scope.lookup2('noSuchGetter').getter,
     );
 
     assertElementNull(
-      scope.lookup(id: 'noSuchSetter', setter: true),
+      scope.lookup2('noSuchSetter').setter,
     );
   }
 
-  test_scope_prefersLocal() async {
+  test_scope_lookup2__prefersLocal() async {
     await assertNoErrorsInCode(r'''
 // ignore:unused_import
 import 'dart:math';
@@ -176,17 +176,17 @@ int sin() => 3;
     var scope = result.libraryElement.scope;
 
     assertElement(
-      scope.lookup(id: 'sin', setter: false),
+      scope.lookup2('sin').getter,
       findElement.topFunction('sin'),
     );
 
     assertElement(
-      scope.lookup(id: 'cos', setter: false),
+      scope.lookup2('cos').getter,
       findElement.importFind('dart:math').topFunction('cos'),
     );
   }
 
-  test_scope_prefix() async {
+  test_scope_lookup2__prefix() async {
     await assertNoErrorsInCode(r'''
 // ignore:unused_import
 import 'dart:math' as math;
@@ -195,12 +195,12 @@ import 'dart:math' as math;
     var scope = result.libraryElement.scope;
 
     assertElement(
-      scope.lookup(id: 'math', setter: false),
+      scope.lookup2('math').getter,
       findElement.prefix('math'),
     );
   }
 
-  test_scope_respectsCombinator_hide() async {
+  test_scope_lookup2__respectsCombinator_hide() async {
     await assertNoErrorsInCode(r'''
 // ignore:unused_import
 import 'dart:math' hide sin;
@@ -209,21 +209,21 @@ import 'dart:math' hide sin;
     var scope = result.libraryElement.scope;
 
     assertElementNull(
-      scope.lookup(id: 'sin', setter: false),
+      scope.lookup2('sin').getter,
     );
 
     var mathFind = findElement.importFind('dart:math');
     assertElement(
-      scope.lookup(id: 'cos', setter: false),
+      scope.lookup2('cos').getter,
       mathFind.topFunction('cos'),
     );
     assertElement(
-      scope.lookup(id: 'tan', setter: false),
+      scope.lookup2('tan').getter,
       mathFind.topFunction('tan'),
     );
   }
 
-  test_scope_respectsCombinator_show() async {
+  test_scope_lookup2__respectsCombinator_show() async {
     await assertNoErrorsInCode(r'''
 // ignore:unused_import
 import 'dart:math' show sin;
@@ -232,11 +232,142 @@ import 'dart:math' show sin;
     var scope = result.libraryElement.scope;
 
     assertElement(
+      scope.lookup2('sin').getter,
+      findElement.importFind('dart:math').topFunction('sin'),
+    );
+
+    assertElementNull(
+      scope.lookup2('cos').getter,
+    );
+  }
+
+  test_scope_lookup_implicitCoreImport() async {
+    await assertNoErrorsInCode('');
+
+    var scope = result.libraryElement.scope;
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'int', setter: false),
+      intElement,
+    );
+  }
+
+  test_scope_lookup_lookup() async {
+    await assertNoErrorsInCode(r'''
+int foo;
+''');
+
+    var scope = result.libraryElement.scope;
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'foo', setter: false),
+      findElement.topGet('foo'),
+    );
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'foo', setter: true),
+      findElement.topSet('foo'),
+    );
+  }
+
+  test_scope_lookup_notFound() async {
+    await assertNoErrorsInCode('');
+
+    var scope = result.libraryElement.scope;
+
+    assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'noSuchGetter', setter: false),
+    );
+
+    assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'noSuchSetter', setter: true),
+    );
+  }
+
+  test_scope_lookup_prefersLocal() async {
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'dart:math';
+
+int sin() => 3;
+''');
+
+    var scope = result.libraryElement.scope;
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'sin', setter: false),
+      findElement.topFunction('sin'),
+    );
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'cos', setter: false),
+      findElement.importFind('dart:math').topFunction('cos'),
+    );
+  }
+
+  test_scope_lookup_prefix() async {
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'dart:math' as math;
+''');
+
+    var scope = result.libraryElement.scope;
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'math', setter: false),
+      findElement.prefix('math'),
+    );
+  }
+
+  test_scope_lookup_respectsCombinator_hide() async {
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'dart:math' hide sin;
+''');
+
+    var scope = result.libraryElement.scope;
+
+    assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'sin', setter: false),
+    );
+
+    var mathFind = findElement.importFind('dart:math');
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'cos', setter: false),
+      mathFind.topFunction('cos'),
+    );
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'tan', setter: false),
+      mathFind.topFunction('tan'),
+    );
+  }
+
+  test_scope_lookup_respectsCombinator_show() async {
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'dart:math' show sin;
+''');
+
+    var scope = result.libraryElement.scope;
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
       scope.lookup(id: 'sin', setter: false),
       findElement.importFind('dart:math').topFunction('sin'),
     );
 
     assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
       scope.lookup(id: 'cos', setter: false),
     );
   }