+++ pkg/analyzer/test/src/dart/resolution/prefix_element_test.dart
@@ -17,7 +17,7 @@ main() {
 
 @reflectiveTest
 class PrefixElementTest extends DriverResolutionTest {
-  test_scope() async {
+  test_scope_lookup() async {
     newFile('/test/lib/a.dart', content: r'''
 var foo = 0;
 ''');
@@ -31,17 +31,43 @@ import 'a.dart' as prefix;
     var importFind = findElement.importFind('package:test/a.dart');
 
     assertElement(
+      // ignore: deprecated_member_use_from_same_package
       scope.lookup(id: 'foo', setter: false),
       importFind.topGet('foo'),
     );
 
     assertElement(
+      // ignore: deprecated_member_use_from_same_package
       scope.lookup(id: 'foo', setter: true),
       importFind.topSet('foo'),
     );
   }
 
-  test_scope_ambiguous_notSdk_both() async {
+  test_scope_lookup2() async {
+    newFile('/test/lib/a.dart', content: r'''
+var foo = 0;
+''');
+
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'a.dart' as prefix;
+''');
+
+    var scope = findElement.prefix('prefix').scope;
+    var importFind = findElement.importFind('package:test/a.dart');
+
+    assertElement(
+      scope.lookup2('foo').getter,
+      importFind.topGet('foo'),
+    );
+
+    assertElement(
+      scope.lookup2('foo').setter,
+      importFind.topSet('foo'),
+    );
+  }
+
+  test_scope_lookup2_ambiguous_notSdk_both() async {
     newFile('/test/lib/a.dart', content: r'''
 var foo = 0;
 ''');
@@ -64,7 +90,7 @@ import 'b.dart' as prefix;
     var bImport = findElement.importFind('package:test/b.dart');
 
     _assertMultiplyDefinedElement(
-      scope.lookup(id: 'foo', setter: false),
+      scope.lookup2('foo').getter,
       [
         aImport.topGet('foo'),
         bImport.topGet('foo'),
@@ -72,7 +98,7 @@ import 'b.dart' as prefix;
     );
 
     _assertMultiplyDefinedElement(
-      scope.lookup(id: 'foo', setter: true),
+      scope.lookup2('foo').setter,
       [
         aImport.topSet('foo'),
         bImport.topSet('foo'),
@@ -80,7 +106,7 @@ import 'b.dart' as prefix;
     );
   }
 
-  test_scope_ambiguous_notSdk_first() async {
+  test_scope_lookup2_ambiguous_notSdk_first() async {
     newFile('/test/lib/a.dart', content: r'''
 var pi = 4;
 ''');
@@ -97,12 +123,12 @@ import 'dart:math' as prefix;
     var aImport = findElement.importFind('package:test/a.dart');
 
     assertElement(
-      scope.lookup(id: 'pi', setter: false),
+      scope.lookup2('pi').getter,
       aImport.topGet('pi'),
     );
   }
 
-  test_scope_ambiguous_notSdk_second() async {
+  test_scope_lookup2_ambiguous_notSdk_second() async {
     newFile('/test/lib/a.dart', content: r'''
 var pi = 4;
 ''');
@@ -119,12 +145,12 @@ import 'a.dart' as prefix;
     var aImport = findElement.importFind('package:test/a.dart');
 
     assertElement(
-      scope.lookup(id: 'pi', setter: false),
+      scope.lookup2('pi').getter,
       aImport.topGet('pi'),
     );
   }
 
-  test_scope_ambiguous_same() async {
+  test_scope_lookup2_ambiguous_same() async {
     newFile('/test/lib/a.dart', content: r'''
 var foo = 0;
 ''');
@@ -145,17 +171,17 @@ import 'b.dart' as prefix;
     var importFind = findElement.importFind('package:test/a.dart');
 
     assertElement(
-      scope.lookup(id: 'foo', setter: false),
+      scope.lookup2('foo').getter,
       importFind.topGet('foo'),
     );
 
     assertElement(
-      scope.lookup(id: 'foo', setter: true),
+      scope.lookup2('foo').setter,
       importFind.topSet('foo'),
     );
   }
 
-  test_scope_differentPrefix() async {
+  test_scope_lookup2_differentPrefix() async {
     newFile('/test/lib/a.dart', content: r'''
 var foo = 0;
 ''');
@@ -176,23 +202,23 @@ import 'b.dart' as prefix2;
     var importFind = findElement.importFind('package:test/a.dart');
 
     assertElement(
-      scope.lookup(id: 'foo', setter: false),
+      scope.lookup2('foo').getter,
       importFind.topGet('foo'),
     );
     assertElement(
-      scope.lookup(id: 'foo', setter: true),
+      scope.lookup2('foo').setter,
       importFind.topSet('foo'),
     );
 
     assertElementNull(
-      scope.lookup(id: 'bar', setter: false),
+      scope.lookup2('bar').getter,
     );
     assertElementNull(
-      scope.lookup(id: 'bar', setter: true),
+      scope.lookup2('bar').setter,
     );
   }
 
-  test_scope_notFound() async {
+  test_scope_lookup2_notFound() async {
     await assertNoErrorsInCode(r'''
 // ignore:unused_import
 import 'dart:math' as math;
@@ -201,15 +227,15 @@ import 'dart:math' as math;
     var scope = findElement.prefix('math').scope;
 
     assertElementNull(
-      scope.lookup(id: 'noSuchGetter', setter: false),
+      scope.lookup2('noSuchGetter').getter,
     );
 
     assertElementNull(
-      scope.lookup(id: 'noSuchSetter', setter: true),
+      scope.lookup2('noSuchSetter').setter,
     );
   }
 
-  test_scope_respectsCombinator_hide() async {
+  test_scope_lookup2_respectsCombinator_hide() async {
     await assertNoErrorsInCode(r'''
 // ignore:unused_import
 import 'dart:math' as math hide sin;
@@ -219,20 +245,20 @@ import 'dart:math' as math hide sin;
     var mathFind = findElement.importFind('dart:math');
 
     assertElementNull(
-      scope.lookup(id: 'sin', setter: false),
+      scope.lookup2('sin').getter,
     );
 
     assertElement(
-      scope.lookup(id: 'cos', setter: false),
+      scope.lookup2('cos').getter,
       mathFind.topFunction('cos'),
     );
     assertElement(
-      scope.lookup(id: 'tan', setter: false),
+      scope.lookup2('tan').getter,
       mathFind.topFunction('tan'),
     );
   }
 
-  test_scope_respectsCombinator_show() async {
+  test_scope_lookup2_respectsCombinator_show() async {
     await assertNoErrorsInCode(r'''
 // ignore:unused_import
 import 'dart:math' as math show sin;
@@ -242,11 +268,238 @@ import 'dart:math' as math show sin;
     var mathFind = findElement.importFind('dart:math');
 
     assertElement(
+      scope.lookup2('sin').getter,
+      mathFind.topFunction('sin'),
+    );
+
+    assertElementNull(
+      scope.lookup2('cos').getter,
+    );
+  }
+
+  test_scope_lookup_ambiguous_notSdk_both() async {
+    newFile('/test/lib/a.dart', content: r'''
+var foo = 0;
+''');
+
+    newFile('/test/lib/b.dart', content: r'''
+var foo = 1.2;
+''');
+
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'a.dart' as prefix;
+
+// ignore:unused_import
+import 'b.dart' as prefix;
+''');
+
+    var scope = findElement.prefix('prefix').scope;
+
+    var aImport = findElement.importFind('package:test/a.dart');
+    var bImport = findElement.importFind('package:test/b.dart');
+
+    _assertMultiplyDefinedElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'foo', setter: false),
+      [
+        aImport.topGet('foo'),
+        bImport.topGet('foo'),
+      ],
+    );
+
+    _assertMultiplyDefinedElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'foo', setter: true),
+      [
+        aImport.topSet('foo'),
+        bImport.topSet('foo'),
+      ],
+    );
+  }
+
+  test_scope_lookup_ambiguous_notSdk_first() async {
+    newFile('/test/lib/a.dart', content: r'''
+var pi = 4;
+''');
+
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'a.dart' as prefix;
+
+// ignore:unused_import
+import 'dart:math' as prefix;
+''');
+
+    var scope = findElement.prefix('prefix').scope;
+    var aImport = findElement.importFind('package:test/a.dart');
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'pi', setter: false),
+      aImport.topGet('pi'),
+    );
+  }
+
+  test_scope_lookup_ambiguous_notSdk_second() async {
+    newFile('/test/lib/a.dart', content: r'''
+var pi = 4;
+''');
+
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'dart:math' as prefix;
+
+// ignore:unused_import
+import 'a.dart' as prefix;
+''');
+
+    var scope = findElement.prefix('prefix').scope;
+    var aImport = findElement.importFind('package:test/a.dart');
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'pi', setter: false),
+      aImport.topGet('pi'),
+    );
+  }
+
+  test_scope_lookup_ambiguous_same() async {
+    newFile('/test/lib/a.dart', content: r'''
+var foo = 0;
+''');
+
+    newFile('/test/lib/b.dart', content: r'''
+export 'a.dart';
+''');
+
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'a.dart' as prefix;
+
+// ignore:unused_import
+import 'b.dart' as prefix;
+''');
+
+    var scope = findElement.prefix('prefix').scope;
+    var importFind = findElement.importFind('package:test/a.dart');
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'foo', setter: false),
+      importFind.topGet('foo'),
+    );
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'foo', setter: true),
+      importFind.topSet('foo'),
+    );
+  }
+
+  test_scope_lookup_differentPrefix() async {
+    newFile('/test/lib/a.dart', content: r'''
+var foo = 0;
+''');
+
+    newFile('/test/lib/b.dart', content: r'''
+var bar = 0;
+''');
+
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'a.dart' as prefix;
+
+// ignore:unused_import
+import 'b.dart' as prefix2;
+''');
+
+    var scope = findElement.prefix('prefix').scope;
+    var importFind = findElement.importFind('package:test/a.dart');
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'foo', setter: false),
+      importFind.topGet('foo'),
+    );
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'foo', setter: true),
+      importFind.topSet('foo'),
+    );
+
+    assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'bar', setter: false),
+    );
+    assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'bar', setter: true),
+    );
+  }
+
+  test_scope_lookup_notFound() async {
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'dart:math' as math;
+''');
+
+    var scope = findElement.prefix('math').scope;
+
+    assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'noSuchGetter', setter: false),
+    );
+
+    assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'noSuchSetter', setter: true),
+    );
+  }
+
+  test_scope_lookup_respectsCombinator_hide() async {
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'dart:math' as math hide sin;
+''');
+
+    var scope = findElement.prefix('math').scope;
+    var mathFind = findElement.importFind('dart:math');
+
+    assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'sin', setter: false),
+    );
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'cos', setter: false),
+      mathFind.topFunction('cos'),
+    );
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
+      scope.lookup(id: 'tan', setter: false),
+      mathFind.topFunction('tan'),
+    );
+  }
+
+  test_scope_lookup_respectsCombinator_show() async {
+    await assertNoErrorsInCode(r'''
+// ignore:unused_import
+import 'dart:math' as math show sin;
+''');
+
+    var scope = findElement.prefix('math').scope;
+    var mathFind = findElement.importFind('dart:math');
+
+    assertElement(
+      // ignore: deprecated_member_use_from_same_package
       scope.lookup(id: 'sin', setter: false),
       mathFind.topFunction('sin'),
     );
 
     assertElementNull(
+      // ignore: deprecated_member_use_from_same_package
       scope.lookup(id: 'cos', setter: false),
     );
   }